% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAggregateData.R
\name{addVariables}
\alias{addVariables}
\title{Add variables from another data set, matching by patient identifier.}
\usage{
addVariables(data, additional.data, subject = "usubjid", vars = "trt")
}
\arguments{
\item{data}{The data to have one or more new variables added}

\item{additional.data}{The dataset containing the new variables to be added to \code{data}}

\item{subject}{The unique subject identifier. Defaults to \code{subject="usubjid"}}

\item{vars}{Character vector with the names of the varialbes in \code{additional.data} to be added to \code{data}}
}
\description{
Add variables from another data set, matching by patient identifier.
}
\details{
The function adds the first value of \code{var} as matched by \code{subject}. If
         that's not unique, you want to subset \code{additional.data} to make it unique,
         or sort it appropriately first.
         An alternative is to use the \code{join} function in the \code{plyr}
         package. NOTE that the call should
         be like \code{mydata <- addVariables(mydata, otherdata, vars="trt")}:
         that is, the returned object is
         a \code{data.frame}, NOT a new column to be appended to \code{myhdata}.
         If there are any factors in the data, unused levels are silently dropped.
}
\examples{
lb <- addVariables(lb, dm, vars="armcd")
}

