% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfpe.R
\name{AIC.lmr}
\alias{AIC.lmr}
\title{Robust version of AIC for an MM-estimated linear model}
\usage{
\method{AIC}{lmr}(object, scale, ..., k = 2)
}
\arguments{
\item{object}{An object of class 'lmr'.}

\item{scale}{An optional scale parameter. If not provided, it is taken from \code{object}.}

\item{...}{Not used.}

\item{k}{The penalty per parameter to be used.}
}
\description{
Robust version of AIC for an MM-estimated linear model
}
\details{
The robust AIC correponds to equation (16) of Tharmaratnam and Claeskens.
    The 'penalty' term is not a simple function of the number of parameters in the
    model, so increaseing k does not necessarily result in simpler models being
    chosen.
}
\references{
Tharmaratnam, K. and Claeskens, G. A comparison of robust versions
     of the AIC based on M, S and MM-estimators, Technical Report KBI 1014,
     Katholieke Universiteit Leuven, 2010.
     https://lirias.kuleuven.be/bitstream/123456789/274771/1/KBI_1014.pdf
}
