% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfpe.R
\name{bisquare}
\alias{bisquare}
\title{Bisquare weight function and its derivatives}
\usage{
bisquare(x, c = 3.443689, d = 0)
}
\arguments{
\item{x}{Scaled residuals from a linear model.}

\item{c}{The tuning constant to the bisquare function.}

\item{d}{The order of derivative of the bisquare function required. Defaulgs
to \code{d=0}, the other allowed values being 1 and 2.}
}
\description{
Bisquare weight function and its derivatives
}

