% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmr.R
\name{boxplot.lmr}
\alias{boxplot.lmr}
\title{Boxplots of scaled residuals, by a factor in the model.}
\usage{
\method{boxplot}{lmr}(
  x,
  by,
  jitter.width = 0.1,
  box = TRUE,
  xlab = "",
  ylab = "Scaled residuals",
  main = "",
  size = 3,
  ptcol = "orange",
  boxcol = "blue",
  boxfill = "light blue",
  ...
)
}
\arguments{
\item{x}{An object of class 'rlm'.}

\item{by}{A character string giving the name of a factor in \code{x}.}

\item{jitter.width}{The amount of jittering to do. Defaults to \code{jitter.width = 0.1}.}

\item{box}{Whether to do the boxplot. If the amount of data is small, it might be better
to just plot the jittered points. Defaults to \code{box=TRUE}.}

\item{xlab}{Horizontal axis label. Defaults to \code{xlab=""}.}

\item{ylab}{Vertical axis label. Defaults to \code{ylab="Scaled residuals"}.}

\item{main}{Main title. Defaults to blank.}

\item{size}{The size of the plot symbols. Defaults to 3.}

\item{ptcol}{The colour of the residuals on the plot. Defaults to \code{"orange"}.}

\item{boxcol}{The colour of the outline of the boxplots. Defaults to \code{"blue"}.}

\item{boxfill}{The fill colour of the boxplots. Defaults to \code{"light blue"}.}

\item{...}{Additional arguments to \code{boxplot}. Not currently used.}
}
\description{
Boxplots of scaled residuals, split by a factor in the model.
}
