% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmTrafficLight.R
\name{evmTrafficLight}
\alias{evmTrafficLight}
\title{Plot predicted return levels of a safety lab variable with some guidance as to
whether it seems safe, potentially suspect, or dangerous}
\usage{
evmTrafficLight(data = NULL, ptcol = "blue", loref = 1/1000,
  hiref = 1/200, refcol = c("green", "orange", "red"),
  linecol = c("blue", "blue"), ptsize = 4, linesize = c(0.5, 1.5),
  scales = "free", ncol = NULL, xlab = "P( > RL)", ylab = "", M,
  main = NULL, ...)
}
\arguments{
\item{data}{An object created by \code{summary.simulate.margarita.prob}}

\item{ptcol}{The colour of points on the plot. Defaults to \code{ptcol="blue"}}

\item{loref}{The lower reference level beneath which a return level is considered
to be acceptably low}

\item{hiref}{The upper reference level above which a return level is considered to
be unacceptably high}

\item{refcol}{Background colours for the acceptable, intermediate, and unaccepable
return levels}

\item{linecol}{Colour of the lines representing the credible intervals on the
predicted return levels}

\item{ptsize}{The size of the point estimate (median) predicted return level}

\item{linesize}{The weight of the lines representing the credible intervals}

\item{scales}{Whether to allow trellis panel scales to match each other or be free}

\item{ncol}{The number of columns in the output plots}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{M}{A vector specifying the return levels used}

\item{main}{The main title}

\item{...}{Other arguments to the plotting functions. Currently unused}
}
\description{
Plot predicted return levels of a safety lab variable with some guidance as to
whether it seems safe, potentially suspect, or dangerous
}
\details{
The function is a fairly simple modification of \code{plot.summary.simulate.margarita.prob}
}
