% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable.sim.R
\name{formattable.summary.margarita.sim.rl}
\alias{formattable.summary.margarita.sim.rl}
\alias{formattable.summary.margarita.sim.prob}
\title{Create an HTML table for outputing via Rmd.}
\usage{
\method{formattable}{summary.margarita.sim.rl}(data, digits = 2,
  backgroundColor = NULL, debug = FALSE)
}
\arguments{
\item{data}{The output of \code{summary.margarita.sim.rl}.}

\item{digits}{The number of digits to round numbers to. Defaults to
\code{digits=2}.}

\item{backgroundColor}{Rows are colored according to treatment group. Defaults
to using \code{shadesOfGrey} to select the correct number of shades of grey.}

\item{debug}{If \code{debug = TRUE}, the list of formatters is returned for
inspection.}
}
\description{
Create an HTML table for outputing via Rmd.
}
\note{
The returned object (if \code{debug = FALSE}) is a formattable object
  and, if desired, can be turned into a \code{DT::datatable} via
  \code{formattable::as.datatable}.
}
