% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAggregateData.R
\name{getAggregateData}
\alias{getAggregateData}
\title{Take an SDTM type lab dataset and return a dataset ready for extreme value modelling}
\usage{
getAggregateData(data, subject = "usubjid", visit = "visitnum",
  baseline.visit = 1, max.study.visit = Inf, baseline = "baseline",
  value = "aval", aggregate.fun = max, also.keep = NULL)
}
\arguments{
\item{data}{A \code{data.frame}, usually the lab data, or vital signs data}

\item{subject}{The unique subject identifier. Defaults to \code{usubjid}}

\item{visit}{The name of the visit variable. This should be numeric rather than character. Defaults to \code{visitnum}}

\item{baseline.visit}{The value of \code{visit} corresponding to the baseline visit. Defaults to \code{visit = 1}}

\item{max.study.visit}{Th evalue of \code{visit} corresponding to the maximum study visit, allowing the user to discard codes representing dropout or follow-up visits of no interest. Defaults to \code{max.study.visit = Inf}}

\item{baseline}{The name of the column containing the baseline values, if any. Defaults to \code{base}}

\item{value}{The name of the column containing the test values. Defaults to \code{aval}}

\item{aggregate.fun}{The function to do the aggregation. Defaults to \code{max}}

\item{also.keep}{The names of any additional variables to keep (like a treatment factor). Defaults
to \code{also.keep=NULL}.}
}
\description{
Take an SDTM type lab dataset and return a dataset ready for extreme value modelling
}
\details{
Subjects with no post-baseline values are excluded from the output data.
         Also, no treatment codes are included in the output; these should be spliced
         in from another dataset.
         Also, if a patient has a missing baseline value or no non-missing post-baseline
         values, they are silently dropped from the returned data object.
         If there are any factors (as a result of specifying \code{also.keep}), unused
         levels are silently dropped.
         The code was commissioned by AstraZeneca and ought to work with their
         implementation of SDTM: other implementations may differ.
}
\examples{
alt <- getBaselines(lb[lb$lbtestcd == "ALT", ], visit="visitnum")
          alt <- getAggregateData(alt, visit="visitnum")
          alt <- addVariables(alt, dm, vars="armcd")
}

