% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAggregateData.R
\name{getBaselines}
\alias{getBaselines}
\title{Add baseline values of the analysis variable to an analysis dataset}
\usage{
getBaselines(data, subject = "usubjid", visit = "visit",
  baseline.visit = 1, value = "aval", baseline = "baseline")
}
\arguments{
\item{data}{A \code{data.frame}}

\item{subject}{The unique patient identifier. Defaults to \code{id = "usubid"}}

\item{visit}{The name of the visit column. Defaultso to \code{visit="visit"}, but
a days on study column could be used instead}

\item{baseline.visit}{The value of the visit column at the time the baselne
measurements are made}

\item{value}{The column containing the data values to be analysed}

\item{baseline}{The name of the baseline column. If it is already in the data and not
empty, the function will exit with a warning. Otherwise, this is the
name of the baseline variable in the output dataset.}
}
\description{
Add baseline values of the analysis variable to an analysis dataset
}
\details{
If multiple measurements are available for some patients at baseline,
         one is chosen almost arbitrarily.
}

