\name{getSegmentData}
\alias{getSegmentData}
\title{Get data for plotting segments representing interval estimates.}
\usage{
getSegmentData(data)
}
\arguments{
  \item{data}{A data frame with either 5 or 3 columns
  representing the point}
}
\description{
Get data for plotting segments representing interval
estimates.
}

