% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAggregateData.R
\name{getStudyDay}
\alias{getStudyDay}
\title{Compute study day (days since randomization, screening, or whatever), from date fields}
\usage{
getStudyDay(data, datecol = "vsdt", subject = "usubjid", visit = "visit",
  baseline.visit = 1)
}
\arguments{
\item{data}{A data.frame}

\item{datecol}{The name of the column containing the dates of the study visits. Defaults to \code{datecol="vsdt"}}

\item{subject}{The name of the column containing the unique subject IDs. Defaults to \code{id="usubjid"}}

\item{visit}{The name of the columne containing the visit identifiers. Defaults to \code{visit="visit"}}

\item{baseline.visit}{The name of the visit taken to be day 0. Defaults to \code{baseline.visit=1}, but could be numeric}
}
\value{
The same data.frame but with a new column called "day0" and another called "studyday"
}
\description{
Compute study day (days since randomization, screening, or whatever), from date fields
}
\details{
If any patient has more than one value at the baseline visit, one is taken almost arbitrarily
}

