% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoHelpers.R
\name{getTransFun}
\alias{getTransFun}
\title{Get a transformation function and its inverse from a character string}
\usage{
getTransFun(x)
}
\arguments{
\item{x}{A character string indicating 'log', 'sqrt' or 'I'.}
}
\value{
A names list with 2 elements, the first of which, 'tfun' is the
        appropriate transformation function; the second of which, 'itfun', is
        the inverse of 'tfun'
}
\description{
Get a transformation function and its inverse from a character string
}

