% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.evm.R
\name{gpdThresh}
\alias{gpdThresh}
\alias{gpdThresh,}
\alias{ggplot.gpdThresh}
\alias{ggplot.mrl}
\alias{ggplot.gpdRangeFit}
\title{Produce plots to aid threshold selection for GPD models}
\usage{
gpdThresh(x, umin = quantile(x, 0.05), umax = quantile(x, 0.95),
  nint = 25, which = 1:2, priorParameters = NULL, cov = "observed")
}
\arguments{
\item{x}{A numeric vector.}

\item{umin}{The minimum value of x to use as a threshold. Defaults to \code{umin=quantile(x, .05)}.}

\item{umax}{The maximum value of x to use as a threshold. Defaults to \code{umin=quantile(x, .95)}.}

\item{nint}{The number of values of x at which to compute the parameters. Defaults to \code{nint=25}.}

\item{which}{Which plots to produce. If \code{which} contains 1, the mean residual life plot is included; if it contains 2, the plots of the generalized Pareto parameters over increasing thresholds are included; if it includes 3, the plot of the EGP3 power parameter is also included. Defaults to including all 3.}

\item{priorParameters}{A list containing the parameters of the prior distribution, passed through to \code{evm}.}

\item{cov}{How to compute the covariance, passed through to \code{evm}. Defaults to \code{cov='observed'}, but you might want to use \code{cov='numeric'} if the computations are unstable.}

\item{...}{Additional arguments passed to \code{ggplot}.}
}
\value{
A list of graphical objects created by \code{ggplot}.
}
\description{
Produce plots to aid threshold selection for GPD models
}
\details{
The use of egp3 was experimental. Experience suggests it's better to think
  carefully about where a sensible threshold might be before using egp3. Note that
  the dimension of the parameter vector for egp3 is greater than that for gp so
  that if you use non-NULL prior parameters, the function can't cope. If you really
  want to use prior parameters, call the functions separately (egp3RangeFit, mrl, gpdRangeFit).
}
\keyword{models}
