% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelp.R
\docType{data}
\name{lb}
\alias{lb}
\alias{dm}
\title{Data included in the margarita package}
\description{
Data included in the margarita package
}
\details{
The \code{liver} data, as found in the texmex (Southworth and Heffernan)
         package, restructured so as to resemble CDISC SDTM datasets. The
         lab values (ALT, AST, TBL and ALP) and dose groups are identical to
         the \code{liver} dataset. The remaining values (age, sex, race and
         so on) are simulated. The data are broken down into the lab data
         (\code{lb}) and (largely fictional) `demography' data (\code{dm}).
         The purpose is to illustrate the usage of some of the functions provided
         by margarita to aid transformation of SDTM data into evm-ready analysis
         data.
}
\references{
Harry Southworth and Janet E. Heffernan (2014). texmex: Statistical
             modelling of extreme values. R package version 2.3.
}
\author{
Harry Southworth \email{harry.southworth@gmail.com}
}
\keyword{data}
