% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmr.R
\name{lmr}
\alias{lmr}
\title{Robust regression using MM-estimation}
\usage{
lmr(formula, data, weights, psi = NULL, method = "MM", c = 3.443689,
  engine = "rlm", maxit = 40, ...)
}
\arguments{
\item{formula}{A formula describing a linear model.}

\item{data}{An appropriate data frame.}

\item{weights}{Not used. This is only here because \code{ggplot2::geom_smooth}
appears to require any custom smoother to take the argument.}

\item{psi}{The psi function to use. The default is to use bisquare weight functions.
Note that if \code{engine = "rlm"} then psi is set to \code{MASS::psi.bisquare}
(i.e. a function) and if \code{engine = "lmrob"}, psi is set to \code{"bisqare"}
(i.e. a character string). See the help files for \code{rlm} and \code{lmrob}
for further information on available values for psi.}

\item{method}{The robust fitting method. Defaults to \code{method="MM"}.}

\item{c}{Tuning parameter to the MM-algorithm. Defaults to \code{c=3.443689} giving 85\% efficiency for Gaussian data.}

\item{engine}{Character string specifying either 'rlm' in which case \code{MASS::rlm} is used,
or 'lmrob' in which case \code{robustbase::lmrob} is used. In the latter case, a robust
version of R^2 is provided, but the default output produces p-values based on t-distributions
that have no theoretical justification. Bootstrapping would be much better.}

\item{maxit}{The maximum number of iterations to perform. Defaults to \code{maxit = 40}}

\item{...}{Other arguments to be passed to \code{rlm} (like contrasts)}
}
\value{
An object of class 'rlm', fit by the function in the MASS package.
        \code{lmr} is just a simple wrapper to \code{rlm}. The returned
        object has an additional component, \code{cov}.

A list containing the same elements as an object of class 'lmr' but with
       additional elements containing the covariance matrix of the parameter
       estimates ('cov'), the data provided in the call to \code{lmr} ('data'),
       the tuning constant for the bisquare loss functions ('c'), and the
       residual degrees of freedom.
}
\description{
Robust regression using MM-estimation with 85\% efficiency for Gaussian data.
}
\details{
The tuning constant for the bisquare function defaults to
       \code{c=3.443689} providing 85\% efficiency for Gaussian data.
       Maronna et al suggest bisquare weight functions and 85\% efficiency
       with MM-estimation in Sections 5.9 and 11.2 of their book. In the setting of
       eliminating baseline effects from clinical trial data, the models
       considered are fairly simple and these defaults appear to work well.
       The value of 3.443689 is 'borrowed' from \code{lmRob} and its support
       functions in the 'robust' package. Rounded values for various Gaussian
       efficiencies appear in Seciton 2.2 of Maronna et al.

       Note that \code{rlm} produces an object
       that inherits from \code{lm}, but \code{lmr} does not in order to avoid
       \code{lm} methods being used on it. Also, \code{lmr} attaches the
       residual degrees of freedom to the object; \code{rlm} deliberately sets
       it to NA to avoid erroneous application of \code{lm} methods.
}
\references{
Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth Edition. Springer, New York. ISBN 0-387-95457-0
            Maronna, R. A, Martin, R. D and Yohai, V. J. (2006) Robust Statistics: Theory and Methods, Wiley
}
\keyword{models}
