% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margarita-package.R
\docType{package}
\name{margarita-package}
\alias{margarita-package}
\title{Combine robust regression and extreme value modelling}
\description{
details here
}
\author{
harry.southworth@gmail.com
}
\keyword{package}

