% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{margarita.getProbs}
\alias{margarita.getProbs}
\title{Get probabilities of threshold exceedance for a GPD model.}
\usage{
margarita.getProbs(X, par, u, p, r, m)
}
\arguments{
\item{X}{An index parameter used in a call to \code{sapply} or \code{lapply}.}

\item{par}{A numberic vector containin the parameters for the GPD model.}

\item{u}{The threshold.}

\item{p}{The rate of threshold exceedance.}

\item{r}{The original data to which the model was fit.}

\item{m}{The threshold above which we wish to estimated the exceedance rate.}
}
\description{
Get probabilities of threshold exceedance for a genralized Pareto model.
}

