% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margarita.R
\name{margarita}
\alias{margarita}
\title{Create an object of class 'margarita'}
\usage{
margarita(
  rlm,
  evmSim,
  newdata = NULL,
  trans = log,
  invtrans = exp,
  alpha = c(0.1, 0.5),
  baseline = "alt.b",
  arm = "arm",
  minima = FALSE
)
}
\arguments{
\item{rlm}{An object of class 'rlm' returned by \code{lmr}.}

\item{evmSim}{An object of class 'evmSim'.}

\item{newdata}{A \code{data.frame} containing the new data from which
predictions are to be made. Defaults to \code{newdata=NULL} which will
only work of there are no covariates in the linear model or the extreme
value model (with the exception of the baseline term in the linear model)}

\item{trans}{A function matching that used to transform the response
in the robust regression. Defaults to \code{trans=log}. Use
\code{trans=I} if no transformation was made.}

\item{invtrans}{A function, the inverse of \code{trans}.}

\item{alpha}{Used to specify coverage of interval estimates. Defaults to
\code{alpha = c(1, .5)}.}

\item{baseline}{Character string giving the name of the baseline variable.}

\item{arm}{Character string giving the name of the treatment group variable.
Defaults to \code{arm = "arm"}. If there is no treatment factor,
presumably there is a continuous exposure variable and the sample
size is the number of rows in the data.}

\item{minima}{Are the extremes minima rather than maxima (i.e. the response
was multiplied by -1 prior to all modelling). Defaults to
\code{minima=FALSE}. Note that if you specify \code{minima=TRUE} the
function will fail and you will be told to multiply by -1 at the very
start of the analysis, and to then multiply M by -1 in the call to
this function.}
}
\description{
Create an object of class 'margarita'
}
\details{
Returns a list with class 'margarita', to be used with
      \code{simulate.margarita}.
}
\keyword{models}
