% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{margarita.rp.matrix}
\alias{margarita.rp.matrix}
\title{Construct a matrix of thresholds whose rate of exceedance we wish to estimate.}
\usage{
margarita.rp.matrix(M, scale, trans, d, baseline)
}
\arguments{
\item{M}{Numeric vector of thresholds.}

\item{scale}{Whether we are interested in the raw values of M or are treating
them as fold-changes from baseline ('proportional') or absolute
differences ('difference')}

\item{trans}{A function for transforming the data, the same as the function
used to transform the data before fitting the original linear model.}

\item{d}{A list of data.frames}

\item{baseline}{The name of the baseline column in d}
}
\description{
Construct a matrix of thresholds whose rate of exceedance we wish to estimate.
}

