% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{margarita.rp}
\alias{margarita.rp}
\title{Return period for GPD}
\usage{
margarita.rp(X, xm, u, par, p, r, family, th)
}
\arguments{
\item{X}{An indexing parameter used in a call to \code{sapply} or \code{lapply}.}

\item{xm}{The threshold above which we wish to estimated the exceedance rate.}

\item{u}{The threshold.}

\item{par}{Matrix of parameter values.}

\item{p}{The rate of threshold exceedance.}

\item{r}{The data to which the original model was fit.}

\item{family}{The family object.}

\item{th}{The threshold used in fitting the model.}
}
\description{
Compute the return period for a generalized Pareto distribution
}
