% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{margaritaScale}
\alias{margaritaScale}
\title{Check an allowed value of the scale argument has been given and reduce it to
its first letter}
\usage{
margaritaScale(s)
}
\arguments{
\item{s}{A character string which should be 'raw', 'proportional' or 'difference'
or an abbreviation of one of those. Only the first letter is returned.}
}
\description{
Check an allowed value of the scale argument has been given and reduce it to
its first letter
}

