% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoHelpers.R
\name{ordinalIndicator}
\alias{ordinalIndicator}
\title{Get the ordinal indicator for an integer}
\usage{
ordinalIndicator(x)
}
\arguments{
\item{x}{An integer}
}
\value{
A character string giving the ordinal indicator for x: 'st' if the last
        digit of x is 1, 'nd' if it is 2, 'rd' if it is 3, 'th' otherwise.
}
\description{
Get the ordinal indicator for an integer
}

