% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoHelpers.R
\name{readAutoInputs}
\alias{readAutoInputs}
\alias{makeAutoEvmStuff}
\title{Read user input from text file in order to set up automated extreme value modelling}
\usage{
readAutoInputs(file, make = TRUE, assign. = TRUE)
}
\arguments{
\item{file}{Character string giving the name of the file.}

\item{make}{Whether or not to make additional objects based on the input file
and system information. Defaults to \code{make=TRUE}.}

\item{assign.}{Whether or not to assign the values to R objects. Defaults to \code{assign.=TRUE}.}
}
\value{
Returns a names list, invisibly, containing the values of the named fields
  in the input file. However, the named items in the list are also assigned to
  position 1 unless the call specifies \code{assign.=FALSE}.
}
\description{
Read user input from text file in order to set up automated extreme value modelling
}
