% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoHelpers.R
\name{readData}
\alias{readData}
\title{Read a data file, figuring out if it is a SAS file or something else}
\usage{
readData(file, type = NULL)
}
\arguments{
\item{file}{The name, including the full path, of the data file}

\item{type}{The type, as indicated by the file extension. At present, must
be either 'sas7bdat', 'csv' or NULL (the default). If NULL, the
function uses the file extension to guess.}
}
\description{
Read a data file, figuring out if it is a SAS file or something else
}
\details{
If the file appears to be a SAS file, the function uses read.sas7bdat
         from Matt Shotwell's sas7bdat package. Currently, compressed SAS
         data files are not supported (but see Shotwell's sas7bdat.parso
         package).
}

