% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable.sim.R
\name{shadesOfGrey}
\alias{shadesOfGrey}
\title{Get a vector of grey colors}
\usage{
shadesOfGrey(n)
}
\arguments{
\item{n}{The number of shades of grey to return.}
}
\description{
Get a vector of grey colors
}
\details{
The darkest value returned will be 'grey51' and the lightest will
  be 'grey100' which is white. (The darkest, black, is 'grey0')
}
