% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simLinear}
\alias{simLinear}
\title{Simulate expected responses from a robust linear model, accounting for
uncertainty in the parameter estimates.}
\usage{
simLinear(
  lmod,
  gmod,
  newdata = NULL,
  baseline = "log(alt.b)",
  rawBaseline = "alt.b",
  invtrans = exp
)
}
\arguments{
\item{lmod}{An object returned by \code{lmr}.}

\item{gmod}{An object of class 'evmSim'.}

\item{newdata}{A data.frame containing values of the covariates used
to fit the robust model, excluding the baseline variable.}

\item{baseline}{Character string identifying the baseline column of the
design matrix used to fit the robust linear model. Defaults to
'log(b.alt)'.}

\item{rawBaseline}{Character string giving the name of the untransformed
baseline values. If no transformation is performed, this will be
the same value as for \code{baseline}. Defaults to
\code{rawBaseline = "alt.b"}.}

\item{invtrans}{A function for back-transforming the baseline data if required.
By default, the function assumes the user worked with log-transformed
baseline data, so assumes \code{invtrans=exp}.}
}
\description{
Simulate expected responses from a robust linear model, accounting for
uncertainty in the parameter estimates.
}
\keyword{datagen}
