% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1.lmr.R
\name{step.lmr}
\alias{step.lmr}
\alias{drop1.lmr}
\title{Stepwise model selection for a robust linear model}
\usage{
step.lmr(
  object,
  scope,
  target = "RFPE",
  direction = "backward",
  trace = TRUE,
  steps = 1000,
  k = 2,
  cores = NULL,
  ...
)
}
\arguments{
\item{object}{An object fit by \code{lmr}.}

\item{scope}{The lowest model to consider.}

\item{target}{Character string giving the method for selecting models. Defaults to "RFPE",
the only current alternative being "AIC".}

\item{direction}{What direction to take steps in. Only "backwards" is implemented.}

\item{trace}{Whether or not to report progress. Defaults to \code{trace=TRUE}.}

\item{steps}{The maximum number of steps to take.}

\item{k}{The penalty factor to be used if AIC is being used as the target. Defaults to \code{k=2}.}

\item{cores}{The number of cores to use when running \code{drop1.lmr} in parallel.
Defaults to \code{cores=NULL} and the function will try to guess how many cores
to use.}

\item{...}{Not used (for compatibility with \code{step}).}
}
\description{
Stepwise model selection for a robust linear model
}
\details{
The function uses robust finite prediction error to decide when to stop
       the selection process. In principle, other approaches could be implemented.
}
