% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoHelpers.R
\name{strstrip}
\alias{strstrip}
\title{Strip leading and trailing whitespace from a character string}
\usage{
strstrip(x, leading = TRUE, trailing = TRUE)
}
\arguments{
\item{x}{A vector of character strings.}

\item{leading}{Whether or not to remove leading whitespace. Defaults to \code{leading=TRUE}.}

\item{trailing}{Whether or not to remove trailing whitespace. Defaults to \code{trailing=TRUE}.}
}
\description{
Strip leading and trailing whitespace from a character string
}
\details{
Code copied from http://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace-in-r,
  2014-11-26.
}

