% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoHelpers.R
\name{tidyCoefNames}
\alias{tidyCoefNames}
\title{Remove the factor name from the levels in the names of model coefficients}
\usage{
tidyCoefNames(x)
}
\arguments{
\item{x}{A model object with an 'xlevels' element}
}
\description{
Remove the factor name from the levels in the names of model coefficients
}
\details{
If there are no factors in the model, the coefficient names are
         returned. Otherwise, the function strips the factor name of each
         coefficient name, for each factor. If two factors have overlapping
         level names, the function will fail.
}

