% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICtab.R
\name{evidence}
\alias{evidence}
\title{Get an xtable object that describes change in deviance in terms of
strength of evidence}
\usage{
evidence()
}
\description{
Get an xtable object that describes change in deviance in terms of
strength of evidence
}

