% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gpdThresh}
\alias{ggplot.gpdRangeFit}
\alias{ggplot.gpdThresh}
\alias{ggplot.mrl}
\alias{gpdThresh}
\alias{gpdThresh,}
\title{Produce plots to aid threshold selection for GPD models}
\usage{
gpdThresh(x, umin = quantile(x, 0.05), umax = quantile(x, 0.95),
  nint = 25, which = 1:3, priorParameters = NULL, cov = "observed")
}
\arguments{
\item{x}{A numeric vector.}

\item{umin}{The minimum value of x to use as a threshold. Defaults to \code{umin=quantile(x, .05)}.}

\item{umax}{The maximum value of x to use as a threshold. Defaults to \code{umin=quantile(x, .95)}.}

\item{nint}{The number of values of x at which to compute the parameters. Defaults to \code{nint=25}.}

\item{which}{Which plots to produce. If \code{which} contains 1, the mean residual life plot is included; if it contains 2, the plots of the generalized Pareto parameters over increasing thresholds are included; if it includes 3, the plot of the EGP3 power parameter is also included. Defaults to including all 3.}

\item{priorParameters}{A list containing the parameters of the prior distribution, passed through to \code{evm}.}

\item{cov}{How to compute the covariance, passed through to \code{evm}. Defaults to \code{cov='observed'}, but you might want to use \code{cov='numeric'} if the computations are unstable.}

\item{...}{Additional arguments passed to \code{ggplot}.}
}
\value{
A list of graphical objects created by \code{ggplot}.
}
\description{
Produce plots to aid threshold selection for GPD models
}
\details{
The only argument is the data, so there is no control over other
   settings used by \code{mrl}, \code{gpdRangeFit} and their plot functions.
   If more control is needed, use those functions directly.
}
\keyword{models}

