% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{margarita.rp}
\alias{margarita.rp}
\title{Return period for GPD}
\usage{
margarita.rp(X, xm, u, phi, xi, p, r)
}
\arguments{
\item{X}{An indexing parameter used in a call to \code{sapply} or \code{lapply}.}

\item{xm}{The threshold above which we wish to estimated the exceedance rate.}

\item{u}{The threshold.}

\item{phi}{The log-scale parameter for the GPD model.}

\item{xi}{The shape parameter for the GPD model.}

\item{p}{The rate of threshold exceedance.}

\item{r}{The data to which the original model was fit.}
}
\description{
Compute the return period for a generalized Pareto distribution
}

