% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfpe.R
\name{RFPE}
\alias{RFPE}
\alias{RFPE.lmRob}
\alias{RFPE.lmr}
\title{Robust final prediction error for a linear model}
\usage{
RFPE(x, scale = NULL)
}
\arguments{
\item{x}{A robust linear model, fit by \code{lmr}.}

\item{scale}{The value of the scale to use. Defaults to \code{scale=NULL} and is take from \code{x}.}
}
\value{
A number representing the robust final prediction error.
}
\description{
Robust final prediction error for a linear model
}
\details{
The definition of robust final prediction error is given in Section 5.12
    of Maronna et al. The function is generic, but so far only methods for objects
    fit by \code{lmr} and \code{lmRob} have been implemented (the latter purely for
    testing). It would be straightforward to implement
    for objects returned by \code{rlm} or \code{lmrob}, for example. Only bisquare
    weight functions are considered.
}
\references{
Maronna, R. A, Martin, R. D and Yohai, V. J. (2006) Robust Statistics: Theory and Methods, Wiley
}

