% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.margarita}
\alias{simulate.margarita}
\alias{simulate.margarita.rl}
\alias{simulate.margarita.prob}
\alias{simulate.margarita.simple}
\alias{as.data.frame.margarita.sim.rl}
\alias{print.margarita.sim.rl}
\alias{head.margarita.sim.rl}
\alias{as.data.frame.margarita.sim.prob}
\alias{print.margarita.sim.prob}
\alias{head.margarita.sim.prob}
\alias{summary.margarita.sim.rl}
\alias{summary.margarita.sim.prob}
\alias{ggplot.summary.margarita.sim.rl}
\alias{ggplot.summary.margarita.sim.prob}
\alias{as.data.frame.summary.margarita.sim.rl}
\alias{print.summary.margarita.sim.rl}
\alias{as.data.frame.summary.margarita.sim.prob}
\alias{print.summary.margarita.sim.prob}
\alias{"/.summary.margarita.sim.rl"}
\alias{"*.summary.margarita.sim.rl"}
\alias{"-.summary.margarita.sim.rl"}
\alias{"+.summary.margarita.sim.rl"}
\title{Simulate return levels, probabilities of threshold exceedences or response vectors
from a robust linear model and an extreme value model}
\usage{
\method{simulate}{margarita}(
  object,
  nsim = 1,
  seed = NULL,
  type = "rl",
  M = NULL,
  scale = "raw",
  Mlabels = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class 'margarita'}

\item{nsim}{Only used when \code{type='simple'}. The number of simulated sets of responses
to produce. If \code{nsim=1}, the default, a vector of simulated responses is returned.
Otherwise, a matrix with \code{nsim} columns, each being a simualted response vector..}

\item{seed}{Used to set the seed for the random number generator. Defaults to \code{seed=NULL}.}

\item{type}{What type of prediction is required: 'rl' for return levels averaged
over baseline; 'prob' for threshold exceedance probabilities averaged
over baseline; 'baserl' for return levels across the observed range of
baselines; 'baseprob' for exceedance probabilities over the observed
range of baselines; or 'simple' for simulated values.
Defaults to \code{type = "rl"}. If \code{type='simple'} is used, the result
is a vector or matrix of simulated response variables. Otherwise, objects of
class 'simulate.margarita.rl' or 'simulate.margarita.prob' which have \code{summary}
functions available..}

\item{M}{The return level to be predicted. Defaults to \code{M=1000}. If
\code{type="prob"}, M should be a vector containing the thresholds
whose probabilities of exceedance the user is interested in, on the
scale of the raw data (typically, something like
\code{M = c(1, 3, 5, 20) * ULN}..}

\item{scale}{The type of prediction being made. Valid values are 'raw',
'proportional' and 'difference'. Due to how the calculations are
performed, \code{scale} needs to be specified in the call to
\code{simulate} if threshold exceedence probabilities are being
predicted, or the call to \code{summary.simulate} if return levels
are being predicted. If \code{scale='raw'} it is assumed that the
values of \code{M} are on the same scale as the response variable
in the GPD model. If \code{scale='proportional'}, the values of
\code{M} are taken to be fold-changes (e.g. 2, 5, 10 times baseline
for return levels, or P(yMax > 2, 5, 10 times yBase)). If
\code{scale='difference'}, absolute changes from baseline are
assumed.}

\item{Mlabels}{Labels to be used in the output. Defaults to \code{Mlabels=NULL}
in which case the function guesses at meaningful labels.}

\item{...}{Other arguments passed to \code{simulate}. \code{simulate.margarita.baseline.prob}
accepts arguments \code{grid.n} for the number of points across the range
of baseline values (defaulting to 25) and \code{baseline.range} for the
range to be simulated over (defaulting to the observed range).}

\item{alpha}{In the corresponding \code{summary} methods, the levels of alpha
for posterior interval estimates. Defaults to \code{alpha=c(0.1, 0.5)}
resulting in 90\% and 50\% interval estimates.}
}
\description{
Simulate return levels and probabilities of threshold exceedences from a
robust linear model and an extreme value model. The procedure is specific to
the situation of extreme value modelling of residuals when predictions on
the scale of the original data are required.
}
\details{
If \code{type="prob"}, the function computes simulated probabilities of
  breaching thresholds \code{M}. These are posterior probabilities, not expected
  proportions. The shape of the distribution will often have a mode at 0. If
  \code{M} is quite low, it might also have a mode at 1. Simple point estimates
  and intervals for such a distribution do not convey its shape. The \code{summary}
  function reports the median and quantiles. The median will often be lower
  than the mean. If you need the expected value, you'll need to do more work.
}
\keyword{datagen}
